Added everything for API, now only Vue

+++ BetPLace/Models/BalanceLog.cs
@@ -5,8 +5,8 @@ namespace BetPlace.Models
     public class BalanceLog
     {
         public int Id { get; set; }
-        public double Change { get; set; }
+        public decimal Change { get; set; }
         public int UserId { get; set; }
-        public double CurrentBalance { get; set; }
+        public decimal CurrentBalance { get; set; }
     }
 }
